+++ jdaviz/core/marks.py
@@ -3,12 +3,14 @@ from bqplot.marks import Lines
 from glue.core import HubListener
 from specutils import Spectrum1D
 
+from jdaviz.core.events import SliceToolStateMessage
+
 
 class BaseSpectrumVerticalLine(Lines, HubListener):
     def __init__(self, viewer, x, **kwargs):
         # we'll store the current units so that we can automatically update the
         # positioning on a change to the x-units
-        self._x_unit = viewer.state.reference_data.get_object().spectral_axis.unit
+        self._x_unit = viewer.state.reference_data.get_object(cls=Spectrum1D).spectral_axis.unit
 
         # the location of the marker will need to update automatically if the
         # underlying data changes (through a unit conversion, for example)
@@ -63,7 +65,7 @@ class SpectralLine(BaseSpectrumVerticalLine):
         # setting redshift will set self.x and enable the obs_value property,
         # but to do that we need x_unit set first (would normally be assigned
         # in the super init)
-        self._x_unit = viewer.state.reference_data.get_object().spectral_axis.unit
+        self._x_unit = viewer.state.reference_data.get_object(cls=Spectrum1D).spectral_axis.unit
         self.redshift = redshift
 
         super().__init__(viewer=viewer, x=self.obs_value, stroke_width=1,
@@ -107,3 +109,91 @@ class SpectralLine(BaseSpectrumVerticalLine):
         # re-compute self.x from current redshift (instead of converting that as well)
         self.redshift = self._redshift
         self._x_unit = new_unit
+
+
+class SliceIndicator(BaseSpectrumVerticalLine, HubListener):
+    """
+    Subclass on bqplot Lines to handle slice/wavelength indicator
+    """
+    def __init__(self, viewer, slice=0, **kwargs):
+        self._active = False
+        self._show_if_inactive = True
+
+        self.slice = slice
+        x_all = viewer.data()[0].spectral_axis
+        # _update_data will set self._x_all, self._x_unit, self.x
+        self._update_data(x_all)
+
+        viewer.session.hub.subscribe(self, SliceToolStateMessage,
+                                     handler=self._on_change_state)
+
+        super().__init__(viewer=viewer,
+                         x=self.x[0],
+                         stroke_width=2,
+                         marker='diamond',
+                         fill='none', close_path=False,
+                         labels=['slice'], labels_visibility='label', **kwargs)
+
+        # default to the initial state of the tool since we can't control if this will
+        # happen before or after the initialization of the tool
+        self._on_change_state({'active': True})
+
+    def _slice_to_x(self, slice=0):
+        if not isinstance(slice, int):
+            raise TypeError(f"slice must be of type int, not {type(slice)}")
+        return self._x_all[slice]
+
+    def _update_colors_opacities(self):
+        # orange (accent) if active, import button blue otherwise (see css in app.vue)
+        if not self._show_if_inactive and not self._active:
+            self.visible = False
+            return
+
+        self.visible = True
+        self.colors = ["#c75109" if self._active else "#007BA1"]
+        self.opacities = [1.0 if self._active else 0.9]
+
+    def _on_change_state(self, msg):
+        if isinstance(msg, dict):
+            changes = msg
+        else:
+            changes = msg.change
+
+        for k, v in changes.items():
+            if k == 'active':
+                self._active = v
+            elif k == 'setting_show_indicator':
+                self._show_if_inactive = v
+            elif k == 'setting_show_wavelength':
+                self.labels_visibility = 'label' if v else 'none'
+
+        self._update_colors_opacities()
+
+    def _update_label(self):
+        self.labels = [f'\u00A0{self.x[0]:0.4e} {self._x_unit}']
+
+    @property
+    def slice(self):
+        return self._slice
+
+    @slice.setter
+    def slice(self, slice):
+        self._slice = slice
+        # if this is within the init, the data may not have been set yet,
+        # in which case we'll just set self._slice for the first time, but
+        # do not need to update self.x or label (yet)
+        if hasattr(self, '_x_all'):
+            x_coord = self._slice_to_x(slice)
+            self.x = [x_coord, x_coord]
+            self._update_label()
+
+    def _update_data(self, x_all):
+        # we want to preserve slice number, so we'll do a bit more than the
+        # default unit-conversion in the base class
+        self._x_all = x_all.value
+        self._x_unit = str(x_all.unit)
+        x_coord = self._slice_to_x(self.slice)
+        self.x = [x_coord, x_coord]
+        if self.labels_visibility == 'label':
+            # update label with new value/unit
+            self._update_label()